Remove utils dictionnary (go back to utils)
