+++ app/helpers/es.py
@@ -1,6 +1,6 @@
 from elasticsearch import helpers as eshelpers, Elasticsearch
 import datetime
-import helpers.utils_dictionnary
+import helpers.utils
 import helpers.logging
 import json
 from pygrok import Grok
@@ -176,15 +176,14 @@ class ES:
     def extract_derived_fields(self, doc_fields):
         derived_fields = dict()
         for field_name, grok_pattern in self.settings.config.items("derivedfields"):
-            if helpers.utils_dictionnary.dict_contains_dotkey(doc_fields, field_name, case_sensitive=False):
+            if helpers.utils.dict_contains_dotkey(doc_fields, field_name, case_sensitive=False):
                 if grok_pattern in self.grok_filters.keys():
                     grok = self.grok_filters[grok_pattern]
                 else:
                     grok = Grok(grok_pattern)
                     self.grok_filters[grok_pattern] = grok
 
-                match_dict = grok.match(helpers.utils_dictionnary.get_dotkey_value(doc_fields, field_name,
-                                                                                   case_sensitive=False))
+                match_dict = grok.match(helpers.utils.get_dotkey_value(doc_fields, field_name, case_sensitive=False))
 
                 if match_dict:
                     for match_dict_k, match_dict_v in match_dict.items():
